+++ clang-tools-extra/clang-tidy/openmp/MissingOrderedDirectiveAfterOrderedClauseCheck.h
@@ -1,0 +1,38 @@
+//===--- MissingOrderedDirectiveAfterOrderedClauseCheck.h - clang-tidy *- C++
+//-*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_OPENMP_MISSINGORDEREDDIRECTIVEAFTERORDEREDCLAUSECHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_OPENMP_MISSINGORDEREDDIRECTIVEAFTERORDEREDCLAUSECHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang::tidy::openmp {
+
+/// FIXME: Write a short description.
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/openmp/missing-ordered-directive-after-ordered-clause.html
+class MissingOrderedDirectiveAfterOrderedClauseCheck : public ClangTidyCheck {
+public:
+  MissingOrderedDirectiveAfterOrderedClauseCheck(StringRef Name,
+                                                 ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+  bool isLanguageVersionSupported(const LangOptions &LangOpts) const override {
+    return LangOpts.OpenMP;
+  }
+  std::optional<TraversalKind> getCheckTraversalKind() const override {
+    return TK_IgnoreUnlessSpelledInSource;
+  }
+};
+
+} // namespace clang::tidy::openmp
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_OPENMP_MISSINGORDEREDDIRECTIVEAFTERORDEREDCLAUSECHECK_H